+++ server/src/main/java/com/epam/aidial/core/server/controller/AdminApplyController.java
@@ -8,13 +8,13 @@ import com.epam.aidial.core.openapi.annotations.ApiSchema;
 import com.epam.aidial.core.server.Proxy;
 import com.epam.aidial.core.server.ProxyContext;
 import com.epam.aidial.core.server.config.MergedConfigStore;
-import com.epam.aidial.core.server.data.config.apply.AdminApplyRequest;
-import com.epam.aidial.core.server.data.config.apply.AdminApplyResponse;
-import com.epam.aidial.core.server.data.config.apply.AdminApplyResult;
-import com.epam.aidial.core.server.data.config.apply.AdminApplyStatus;
-import com.epam.aidial.core.server.data.config.apply.AdminManifest;
-import com.epam.aidial.core.server.data.config.apply.ValidationResult;
-import com.epam.aidial.core.server.data.config.apply.ValidationStatus;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplyRequest;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplyResponse;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplyResult;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplyStatus;
+import com.epam.aidial.core.server.data.config.manifest.AdminManifest;
+import com.epam.aidial.core.server.data.config.manifest.ValidationResult;
+import com.epam.aidial.core.server.data.config.manifest.ValidationStatus;
 import com.epam.aidial.core.server.security.ConfigAuthorizationService;
 import com.epam.aidial.core.server.service.config.ConfigApplyService;
 import com.epam.aidial.core.server.service.config.ConfigManifestSupport;
