+++ server/src/main/java/com/epam/aidial/core/server/controller/AdminValidateController.java
@@ -8,11 +8,11 @@ import com.epam.aidial.core.openapi.annotations.ApiSchema;
 import com.epam.aidial.core.server.Proxy;
 import com.epam.aidial.core.server.ProxyContext;
 import com.epam.aidial.core.server.config.MergedConfigStore;
-import com.epam.aidial.core.server.data.config.apply.AdminApplyRequest;
-import com.epam.aidial.core.server.data.config.apply.AdminManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminValidateResponse;
-import com.epam.aidial.core.server.data.config.apply.ValidationResult;
-import com.epam.aidial.core.server.data.config.apply.ValidationStatus;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplyRequest;
+import com.epam.aidial.core.server.data.config.manifest.AdminManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminValidateResponse;
+import com.epam.aidial.core.server.data.config.manifest.ValidationResult;
+import com.epam.aidial.core.server.data.config.manifest.ValidationStatus;
 import com.epam.aidial.core.server.security.ConfigAuthorizationService;
 import com.epam.aidial.core.server.service.config.ConfigManifestSupport;
 import com.epam.aidial.core.server.service.config.ConfigValidationService;
