+++ server/src/main/java/com/epam/aidial/core/server/controller/ConfigFileMigrateController.java
@@ -11,10 +11,10 @@ import com.epam.aidial.core.server.Proxy;
 import com.epam.aidial.core.server.ProxyContext;
 import com.epam.aidial.core.server.config.MergedConfigStore;
 import com.epam.aidial.core.server.config.SchemaMigrationNameResolver;
-import com.epam.aidial.core.server.data.config.apply.AdminApplyStatus;
-import com.epam.aidial.core.server.data.config.apply.AdminManifest;
-import com.epam.aidial.core.server.data.config.apply.ValidationResult;
-import com.epam.aidial.core.server.data.config.apply.ValidationStatus;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplyStatus;
+import com.epam.aidial.core.server.data.config.manifest.AdminManifest;
+import com.epam.aidial.core.server.data.config.manifest.ValidationResult;
+import com.epam.aidial.core.server.data.config.manifest.ValidationStatus;
 import com.epam.aidial.core.server.data.config.migration.ConfigFileMigrateRequest;
 import com.epam.aidial.core.server.data.config.migration.ConfigFileMigrateResponse;
 import com.epam.aidial.core.server.data.config.migration.ConfigFileMigrateResult;
