+++ server/src/main/java/com/epam/aidial/core/server/data/config/manifest/AdminApplicationManifest.java
@@ -1,4 +1,4 @@
-package com.epam.aidial.core.server.data.config.apply;
+package com.epam.aidial.core.server.data.config.manifest;
 
 import com.epam.aidial.core.config.Application;
