+++ server/src/main/java/com/epam/aidial/core/server/data/config/manifest/AdminApplyResult.java
@@ -1,3 +1,3 @@
-package com.epam.aidial.core.server.data.config.apply;
+package com.epam.aidial.core.server.data.config.manifest;
 
 public record AdminApplyResult(String entityId, AdminApplyStatus status, String error) {}
