+++ server/src/main/java/com/epam/aidial/core/server/data/config/manifest/ValidationResult.java
@@ -1,3 +1,3 @@
-package com.epam.aidial.core.server.data.config.apply;
+package com.epam.aidial.core.server.data.config.manifest;
 
 public record ValidationResult(String entityId, ValidationStatus status, String error) {}
