+++ server/src/main/java/com/epam/aidial/core/server/service/config/ConfigManifestSupport.java
@@ -13,19 +13,19 @@ import com.epam.aidial.core.config.Translator;
 import com.epam.aidial.core.server.config.ConfigPostProcessor;
 import com.epam.aidial.core.server.config.MergedConfigStore;
 import com.epam.aidial.core.server.config.ValidationWarning;
-import com.epam.aidial.core.server.data.config.apply.AdminApplicationManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminCatalogSchemaManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminInterceptorManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminKeyManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminModelManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminRoleManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminRouteManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminSchemaManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminSettingsManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminToolSetManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminTranslatorManifest;
-import com.epam.aidial.core.server.data.config.apply.AdminTypedManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminApplicationManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminCatalogSchemaManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminInterceptorManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminKeyManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminModelManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminRoleManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminRouteManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminSchemaManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminSettingsManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminToolSetManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminTranslatorManifest;
+import com.epam.aidial.core.server.data.config.manifest.AdminTypedManifest;
 import com.epam.aidial.core.server.util.ResourceDescriptorFactory;
 import com.epam.aidial.core.storage.resource.ResourceDescriptor;
 import com.epam.aidial.core.storage.resource.ResourceTypes;
